auth: Send an extra initial_req if an authgroup dropdown is present
In XML POST mode, the server will often refuse to take authgroup changes
in the same POST operation as the login credentials[1].  Instead, an
extra xmlpost_initial_req operation is needed to specify which group to
use, and then the server will re-render a new login form with a new
<opaque> section.

So, if XML POST mode is being used, and an authgroup selection is
required, first present the user with a form that only shows the
authgroup options.  After that is submitted (generating the new
initial_req and updating the <opaque> data), prompt for the login and
password as usual.

This won't affect non-XML-POST servers so it should not break anything
that works today.

[1] http://lists.infradead.org/pipermail/openconnect-devel/2013-June/001081.html

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>