android: Fix broken ARCH=mips build
The MIPS NDK doesn't seem to handle dependencies correctly.  i.e.

libopenconnect.so references symbols in libz.so (from the toolchain's
sysroot), so it has a NEEDED entry for libz.so.  It was linked with -lz.

openconnect links in libopenconnect.so, but doesn't use zlib directly so
it isn't linked with -lz.

This causes a link failure on MIPS only:

    /opt/android-ndk-r9/toolchains/mipsel-linux-android-4.8/prebuilt/linux-x86_64/bin/../lib/gcc/mipsel-linux-android/4.8/../../../../mipsel-linux-android/bin/ld: warning: libz.so, needed by ./.libs/libopenconnect.so, not found (try using -rpath or -rpath-link)
    ./.libs/libopenconnect.so: undefined reference to `deflateEnd'
    ./.libs/libopenconnect.so: undefined reference to `inflate'
    ./.libs/libopenconnect.so: undefined reference to `inflateInit2_'
    ./.libs/libopenconnect.so: undefined reference to `deflate'
    ./.libs/libopenconnect.so: undefined reference to `adler32'
    ./.libs/libopenconnect.so: undefined reference to `inflateEnd'
    ./.libs/libopenconnect.so: undefined reference to `deflateInit2_'
    collect2: error: ld returned 1 exit status
    make[2]: *** [openconnect] Error 1

ARM/x86 are fine.  So we'll use a temporary workaround on MIPS only,
until it's resolved.

Reference: https://code.google.com/p/android/issues/detail?id=64266

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>