main: --cookie-on-stdin doesn't work with ocserv cookies
ocserv's cookie strings are typically much longer (512 characters) than
the cookies generated by a Cisco ASA (<100 characters).  So when
network-manager-openconnect tries to connect to ocserv, the connection
will abruptly drop, and this will show up in syslog:

    NetworkManager[717]: <info> VPN plugin state changed: starting (3)
    NetworkManager[717]: <info> VPN connection 'ocserv' (Connect) reply received.
    openconnect[23054]: Attempting to connect to server 192.168.0.1:443
    openconnect[23054]: SSL negotiation with 192.168.0.1
    openconnect[23054]: Connected to HTTPS on 192.168.0.1
    openconnect[23054]: VPN service unavailable; reason: <unknown>

Quick fix is to allow up to 1024 bytes + NUL termination for the cookie
string.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>